Updated documentation and scripts for v 1.1.0.0
Version 1.1.0.0 fixes the layout of bulleted lists and adds friendly
design-time errors for quicker debugging. In the event of an error in
design mode, the error message, the HTML input and the generated XAML
output are all displayed in the control within the designer,  and other
controls on the page are still rendered normally.